refactor: Extract modules from gdal/dataset to reduce class size
Relates to gh-8.
